Deploy na main 12/07/2021

+++ desafio_spring/src/main/java/com/desafiospringboot/Entities/UserClient.java
@@ -2,8 +2,6 @@ package com.desafiospringboot.Entities;
 
 import javax.persistence.Entity;
 import javax.persistence.ManyToMany;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
 import java.util.List;
 
 @Entity
@@ -11,7 +9,7 @@ public class UserClient extends User{
     @ManyToMany
     private List<UserSeller> following;
 
-    public UserClient(long id, String name, List<UserSeller> following) {
+    public UserClient(int id, String name, List<UserSeller> following) {
         super(id, name);
         this.following = following;
     }
