+++ desafio_spring/src/main/java/com/desafiospringboot/Repositories/UserClientRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserClientRepository extends JpaRepository<UserClient, Long> {
+public interface UserClientRepository extends JpaRepository<UserClient, Integer> {
 }
