+++ desafio_spring/src/main/java/com/desafiospringboot/Repositories/UserSellerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserSellerRepository extends JpaRepository<UserSeller, Long> {
+public interface UserSellerRepository extends JpaRepository<UserSeller, Integer> {
 }
